merged_http2/requirements.txt
psycopg2-binary==2.9.9
httpx==0.26.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
mcp==0.1.0
pydantic==2.5.0